make unionfind cleaner find()
control flow seems much cleaner this way.
Also depending on how much the except clause is followed (more during the earlier phase, less during later phase) this can be faster,
or neutral in the amortized worse case.
This is because previously we have been "paying for" the if x not in self.parent[x] all the time regardless.

also try/except improvements in 3.11+ might bring even more enhancement in speed